+++ plone/app/upgrade/__init__.py
@@ -1,11 +1,11 @@
+from importlib.metadata import distribution
+from importlib.metadata import PackageNotFoundError
 from plone.app.upgrade.utils import alias_module
-
-import pkg_resources
 
 
 try:
-    pkg_resources.get_distribution("Products.CMFQuickInstallerTool")
-except pkg_resources.DistributionNotFound:
+    distribution("Products.CMFQuickInstallerTool")
+except PackageNotFoundError:
     # The alias module helps when migrating to Plone 6.0.0a1.
     # Remove when we no longer support upgrading from Plone 5.2.
     from . import bbb_qi