+++ plone/app/upgrade/tests/test_upgrade.py
@@ -67,10 +67,51 @@ class TestUpgrade(MigrationTest):
         upgrades = self.setup.listUpgrades(_DEFAULT_PROFILE)
         self.assertEqual(len(upgrades), 0)
 
+    def test_upgrade_tinymce_menubar(self):
+
+        from packaging.version import parse
+        from plone.app.upgrade.utils import plone_version
+        from plone.app.upgrade.v61.final import upgrade_registry_tinymce_menubar
+        from plone.base.interfaces.controlpanel import ITinyMCESchema
+        from plone.registry import field
+        from plone.registry.interfaces import IRegistry
+        from plone.registry.record import Record
+        from zope.component import getUtility
+
+        version = parse(plone_version)
+        if version.major < 6 or version.minor < 1:
+            # don't execute the test for Plone < 6.1
+            return
+
+        registry = getUtility(IRegistry)
+
+        # for test we delete the original record
+        del registry.records["plone.menubar"]
+
+        # and set a new dummy list record
+        record = registry.records["plone.menubar"] = Record(
+            field.List(
+                title="Test",
+                min_length=0,
+                max_length=10,
+                value_type=field.TextLine(title="Value"),
+            ),
+            ["Test1", "Test2"],
+        )
+
+        # run the upgrade helper function
+        upgrade_registry_tinymce_menubar(self.setup)
+
+        # the interface of the record should be `ITinyMCESchema`
+        self.assertTrue(record.interface, ITinyMCESchema)
+
+        # the value should be a string, not a list
+        self.assertTrue(record.value, "Test1 Test2")
+
 
 def test_suite():
     import unittest
 
-    return unittest.TestSuite((
-        unittest.defaultTestLoader.loadTestsFromTestCase(TestUpgrade),
-    ))
+    return unittest.TestSuite(
+        (unittest.defaultTestLoader.loadTestsFromTestCase(TestUpgrade),)
+    )