+++ plone/app/upgrade/v52/tests.py
@@ -1,12 +1,10 @@
 from DateTime import DateTime
-from pkg_resources import get_distribution
-from pkg_resources import parse_version
 from plone.app.testing import PLONE_INTEGRATION_TESTING
+from plone.base.interfaces import IMarkupSchema
 from plone.registry import field
 from plone.registry import Record
 from plone.registry.interfaces import IRegistry
 from Products.CMFCore.utils import getToolByName
-from plone.base.interfaces import IMarkupSchema
 from zope.component import getUtility
 
 import unittest
@@ -247,14 +245,15 @@ class UpgradePortalTransforms521to522Test(unittest.TestCase):
 
 
 def test_suite():
-    # Skip these tests on Plone < 5.2a1
-    plone_version = get_distribution("Products.CMFPlone").version
-    if not parse_version(plone_version) >= parse_version("5.2a1"):
-        return unittest.TestSuite()
-
-    return unittest.TestSuite((
-        unittest.defaultTestLoader.loadTestsFromTestCase(UpgradeMemberData51to52Test),
-        unittest.defaultTestLoader.loadTestsFromTestCase(Various52Test),
-        unittest.defaultTestLoader.loadTestsFromTestCase(SiteLogoTest),
-        unittest.defaultTestLoader.loadTestsFromTestCase(UpgradePortalTransforms521to522Test),
-    ))
+    return unittest.TestSuite(
+        (
+            unittest.defaultTestLoader.loadTestsFromTestCase(
+                UpgradeMemberData51to52Test
+            ),
+            unittest.defaultTestLoader.loadTestsFromTestCase(Various52Test),
+            unittest.defaultTestLoader.loadTestsFromTestCase(SiteLogoTest),
+            unittest.defaultTestLoader.loadTestsFromTestCase(
+                UpgradePortalTransforms521to522Test
+            ),
+        )
+    )